# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM API Documentation",
        "description": "api crm amirrashidy\n * created by mohammad rezaei ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "لیست محصولات فعال",
                "description": "برمی‌گرداند لیست همه محصولات که فیلد enabled = 1 دارند",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "لیست محصولات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "افزودن محصول جدید",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "qtitle",
                                    "quantity",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "qtitle": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "number"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "information": {
                                        "type": "string"
                                    },
                                    "discount": {
                                        "type": "number"
                                    },
                                    "discount_price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "محصول با موفقیت ثبت شد"
                    },
                    "422": {
                        "description": "داده‌های نامعتبر"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "دریافت اطلاعات یک محصول",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه محصول",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "اطلاعات محصول",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "محصول پیدا نشد"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "ویرایش اطلاعات محصول",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "qtitle",
                                    "quantity",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "qtitle": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "number"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "information": {
                                        "type": "string"
                                    },
                                    "discount": {
                                        "type": "number"
                                    },
                                    "discount_price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "محصول بروزرسانی شد"
                    },
                    "404": {
                        "description": "محصول پیدا نشد"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "حذف یا غیرفعالسازی محصول",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "محصول غیرفعال شد"
                    },
                    "404": {
                        "description": "محصول پیدا نشد"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "qtitle": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "price": {
                        "type": "number"
                    },
                    "discount": {
                        "type": "number"
                    },
                    "discount_price": {
                        "type": "number"
                    },
                    "information": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}